+++ src/main/java/com/homeybites/services/CategoryService.java
@@ -10,14 +10,14 @@ public interface CategoryService {
 	Category addCategory(Category category);
 	
 	//get category
-	Category getCategory(Integer categoryId);
+	Category getCategory(Long categoryId);
 	
 	//get all category
 	List<Category> getAllCategory();
 	
 	//update category
-	Category updateCategory(Category category, Integer categoryId);
+	Category updateCategory(Category category, Long categoryId);
 	
 	//delete category
-	void deleteCategory(Integer categoryId);
+	void deleteCategory(Long categoryId);
 }
